DP-23507: Moving breadcrumb preview code to the mass_hierarchy module and fixed bugs
