+++ docroot/modules/custom/mass_hierarchy/mass_hierarchy.info.yml
@@ -1,6 +1,9 @@
 name: Mass.gov Hierarchy
 type: module
-description: Related to hierarchies, parents, children, etc.
+description: Related to entity hierarchies, parents, children, etc.
 core: 8.x
 core_version_requirement: ^8 || ^9
 package: Custom
+dependencies:
+  - entity_hierarchy
+  - entity_hierarchy_breadcrumb
