+++ docroot/modules/custom/mass_utility/src/MassUtilityServiceProvider.php
@@ -21,11 +21,6 @@ class MassUtilityServiceProvider extends ServiceProviderBase {
       $container->setParameter('monolog.processors', $processors);
     }
 
-    // Override entity_hierarchy.breadcrumb service.
-    if ($container->hasDefinition('entity_hierarchy.breadcrumb')) {
-      $definition = $container->getDefinition('entity_hierarchy.breadcrumb');
-      $definition->setClass('Drupal\mass_utility\MassUtilityBreadcrumb');
-    }
   }
 
 }
